customized_card-Add textButton for each row.When you click on the line, a new window of the volunteer details appears in the line.In addition, I added a context variable to a customized card so that we can use this variable in a navigator function that we use to replace windows.
reports_win-Add context to CustomizedCard.

volunteer_details- A class that inherits from a card  and designs the page of the details of the relevant volunteer This class receives a number of variables such as first name, last name, city, beginning and end of insurance and more ...

details_win-A window which contains the class of volunteer_details.

Bugs:
- Fix the height between each row.
- More beautiful design of the card
- Link the information from the server into volunteer_details.